requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0

# PyTorch and PyTorch Geometric
torch>=1.10.0
torch-scatter>=2.0.9
torch-sparse>=0.6.12
torch-cluster>=1.5.9
torch-spline-conv>=1.2.1
torch-geometric>=2.0.0

# Graph processing
networkx>=2.6.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Logging and monitoring
tensorboard>=2.6.0
wandb>=0.12.0

# Code quality
black>=21.9b0
flake8>=3.9.0
isort>=5.9.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0